DM-7802 LSST time series, part 1 (search processor, webplot creator)
